refactor : 테스트 yml 파일 수정
