+++ src/test/java/com/fourthread/ozang/module/domain/feed/integration/FeedCommentServiceIntegrationTest.java
@@ -41,15 +41,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 @SpringBootTest
 @Transactional
-@TestPropertySource(properties = {
-    "spring.jpa.hibernate.ddl-auto=create-drop",
-    "ADMIN_USERNAME=test-admin",
-    "ADMIN_EMAIL=test-admin@mail.com",
-    "ADMIN_PASSWORD=test-pass",
-    "JWT_SECRET=d12d12d21d21d12d2",
-    "KAKAO_API_KEY=test",
-    "WEATHER_API_KEY=dwqqdd11"
-})
 class FeedCommentServiceIntegrationTest {
 
   @Autowired
