+++ src/test/java/com/fourthread/ozang/module/domain/user/service/UserServiceIntegrationTest.java
@@ -22,21 +22,9 @@ import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.transaction.annotation.Transactional;
 import software.amazon.awssdk.services.s3.S3Client;
 
-@SpringBootTest(properties = {
-    "KAKAO_API_KEY=kakao-test",
-    "WEATHER_API_KEY=api-test",
-    "ADMIN_USERNAME=admin",
-    "ADMIN_EMAIL=admin@mail.com",
-    "ADMIN_PASSWORD=1234",
-    "JWT_SECRET=1dfadfafafvdfa",
-    "cloud.aws.s3.bucket=test-bucket"
-
-})
+@SpringBootTest
 @ActiveProfiles("test")
 public class UserServiceIntegrationTest {
-
-  @MockitoBean
-  private S3Client s3Client;
 
   @Autowired
   private UserService userService;
